% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abstract_build}
\alias{abstract_build}
\title{Build abstract from inverted index}
\usage{
abstract_build(ab, build = TRUE)
}
\arguments{
\item{ab}{List. Inverted index of abstract.}

\item{build}{Logical. If TRUE, build the abstract.}
}
\value{
Character string. The abstract of the paper.
}
\description{
Build abstract from inverted index
}
\keyword{internal}
