% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_author.R
\name{clean_author_info}
\alias{clean_author_info}
\title{Clean Author Info}
\usage{
clean_author_info(author_info)
}
\arguments{
\item{author_info}{list of results from find_author}
}
\value{
tibble of results
}
\description{
Clean Author Info
}
