% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_venue.R
\name{clean_venue_info}
\alias{clean_venue_info}
\title{Clean Venue Info}
\usage{
clean_venue_info(venue_info)
}
\arguments{
\item{venue_info}{list from find_venue function}
}
\value{
tibble with venue information
}
\description{
Clean Venue Info
}
