% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papers_info.R
\name{clean_works_result}
\alias{clean_works_result}
\title{Clean Works Info}
\usage{
clean_works_result(paper_info)
}
\arguments{
\item{paper_info}{list object with results from the find_work call}
}
\value{
dataframe with one row per work object
}
\description{
Clean Works Info
}
