% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concept_abbrev}
\alias{concept_abbrev}
\title{Concepts and abbreviations.}
\format{
A data frame with 19 observations and 3 variables:
\code{id}, \code{display_name}, and \code{abbreviation}.
}
\usage{
concept_abbrev
}
\description{
0-level concepts and corresponding abbreviations.
}
\keyword{datasets}
