% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_author.R
\name{find_author}
\alias{find_author}
\title{Find Author(s)}
\usage{
find_author(
  id_type = c("openalex", "orcid", "mag", "twitter", "wikipedia", "scopus"),
  id
)
}
\arguments{
\item{id_type}{the type of ids you are submitting, should be: openalex, orcid,
mag, twitter, wikipedia, or scopus}

\item{id}{one or more ids}
}
\value{
tibble of results from openalex api
}
\description{
Find Author(s)
}
\examples{
find_author(id_type = "openalex", id = c("A2941293769", "A2296125463"))
}
