% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_concept.R
\name{find_concept}
\alias{find_concept}
\title{Find Concept}
\usage{
find_concept(
  id_type = c("openalex", "wikidata", "mag", "wikipedia", "umls_aui", "umls_cui"),
  id
)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Find Concept
}
