% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_instution.R
\name{find_institution}
\alias{find_institution}
\title{Find Institution}
\usage{
find_institution(
  id_type = c("openalex", "ror", "mag", "grid", "wikipedia", "wikidata"),
  id
)
}
\arguments{
\item{id_type}{the type of ids, should be one of: openalex, ror, mag, grid,
wikipedia, or wikidata}

\item{id}{one or more ids for finding}
}
\value{
tibble of institution info
}
\description{
Find Institution
}
\examples{
find_institution(id_type = "openalex", id = c("I170897317", "I114027177"))
}
