% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papers_info.R
\name{find_paper}
\alias{find_paper}
\title{Find Paper}
\usage{
find_paper(id_type = c("open_alex", "doi"), id)
}
\arguments{
\item{id_type}{the type of id you are using to locate the author: must be either orcid
or openalex}

\item{id}{the id for the author that matches the type selected initially}
}
\value{
list (sorry) of paper info
}
\description{
Get Paper details
}
\examples{
find_paper(id_type = "doi", id = "https://doi.org/10.7717/peerj.4375")
}
