% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_venue.R
\name{find_venue}
\alias{find_venue}
\title{Find Venue}
\usage{
find_venue(id_type = c("openalex", "issn_l", "mag", "issn"), id)
}
\arguments{
\item{id_type}{type of id you are submitting, should be one of openalex,
issn_l, mag, or issn}

\item{id}{one or more ids that correspond to the id type}
}
\value{
tibble of results from api
}
\description{
Find Venue
}
\examples{
find_venue(id_type = "openalex", id = c("https://openalex.org/V131591925",
"https://openalex.org/V125754415"))
}
