% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papers_info.R
\name{find_work}
\alias{find_work}
\title{Find Work}
\usage{
find_work(
  id_type = c("openalex", "doi", "mag", "pmid", "pmcid"),
  id,
  variable_unnest = NULL
)
}
\arguments{
\item{id_type}{the type of id you are using to locate the author: must be either orcid
or openalex}

\item{id}{the id(s) for the work that matches the type selected initially}

\item{variable_unnest}{the column(s) you want to unnest by to make the dataframe longer -
options are authors, concepts, referenced_works, related_works, and
citation_counts_year}
}
\value{
tibble of paper info
}
\description{
Find Work details
}
\examples{
find_work(id_type = "doi", id = "https://doi.org/10.7717/peerj.4375")

}
