% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundational_functions.R
\name{get_all_data_for_query}
\alias{get_all_data_for_query}
\title{Get all data for Query}
\usage{
get_all_data_for_query(base_page_path, number_of_pages)
}
\arguments{
\item{base_page_path}{the string corresponding to the API query}

\item{number_of_pages}{total number of pages of results retured by the API query}
}
\value{
dataframe of results matching your query
}
\description{
Queries the openalex api for the data matching your query
}
