% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors_papers.R
\name{get_authors_papers}
\alias{get_authors_papers}
\title{Get Authors Papers}
\usage{
get_authors_papers(id_type = c("orcid", "openalex", "scopus", "mag"), id)
}
\arguments{
\item{id_type}{the type of id you are using to locate the author: must be either orcid
or openalex}

\item{id}{the id for the author that matches the type selected initially}
}
\value{
dataframe with all papers from that author, includes variables:
id: openalex id of the paper
display_name: title of the paper
publication date: date of publication in YYYY-MM-DD form
relevance score: ?Not sure what this is, all NA for me atm
authorships: nested dataframe of all authors in the paper (includes author.id,
author_position, author.display_name, and author.orcid)
concepts: nested dataframe of the concepts associated with the paper (includes openalex's
id of the concept, display.name of the concept, score of the concept, level of the
concept in the concept hierarchy, and wikidata of the concept)
alternate_host_venues: ?Not sure what this is either, will need to check on openalex's page
cited_by_count: number of other papers that have cited the paper
publication_year: int
cited_by_api_url: openalex api url for the papers that have cited this paper
doi: doi of the paper
type: type of paper (e.g. proceedings-article, journal-article, book-chapter,
posted-content)
is_paratext: ?Not sure what this is either, will need to check on openalex's page
is_retracted: TRUE/FALSE whether paper has been retracted
referenced_works: list of all works referenced in paper (with openalex urls :) )
related_works: similar works? not sure how these are found
}
\description{
Get Authors Papers
}
\examples{
get_authors_papers(id_type = "openalex", id = "A2941293769")
}
