% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coauthors.R
\name{get_coauthors}
\alias{get_coauthors}
\title{Get Coauthors}
\usage{
get_coauthors(id_type = c("orcid", "openalex", "scopus", "mag"), id)
}
\arguments{
\item{id_type}{the type of id for the ego author you are using, should be one of: orcid, openalex, scopus, or mag}

\item{id}{the id matching the author you want to retrieve the coauthors for}
}
\value{
a tibble with 5 columns: paper_id, ego_author_id, ego_author_name, alter_author_id, and alter_author_name
}
\description{
Get Coauthors
}
\examples{
get_coauthors(id_type = "openalex", id = "A2941293769")
}
