% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundational_functions.R
\name{get_data_from_page}
\alias{get_data_from_page}
\title{Get Data from Page}
\usage{
get_data_from_page(page_path)
}
\arguments{
\item{page_path}{}
}
\value{

}
\description{
Convenience function to get the data from each page -- only pulling out the
results
}
\examples{
get_data_from_page("https://api.openalex.org/concepts?filter=level:1")
}
