% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundational_functions.R
\name{get_links_for_each_page}
\alias{get_links_for_each_page}
\title{Get links for each page}
\usage{
get_links_for_each_page(page1_path)
}
\arguments{
\item{page1_path}{the url of the first page of results}
}
\value{
the links to request
}
\description{
Get links for each page
}
