% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundational_functions.R
\name{get_number_of_pages}
\alias{get_number_of_pages}
\title{Get number of pages of results}
\usage{
get_number_of_pages(page1_path)
}
\arguments{
\item{page1_path}{the url of the first page of results}
}
\value{
the total number of pages needed to get all of the results
}
\description{
Get number of pages of results
}
\examples{
get_number_of_pages("https://api.openalex.org/concepts?filter=level:1")
}
