% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oa2df_coverage}
\alias{oa2df_coverage}
\title{Coverage of OpenAlex entity fields after converting to data frame.}
\format{
Each element is a dataframe with 3 columns
 \describe{
    \item{original}{original field name from OpenAlex}
    \item{oa2df}{new column name in output dataframe from oa2df}
    \item{comment}{additional notes}
    }
}
\usage{
oa2df_coverage
}
\description{
List with 8 elements associated with 8 OpenAlex entities.
}
\keyword{datasets}
