% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaApiRequest.R
\name{oa_entities}
\alias{oa_entities}
\title{Available entities in the OpenAlex database}
\usage{
oa_entities()
}
\value{
Character vector of 5 entity options.
}
\description{
Available entities in the OpenAlex database
}
