% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coro.R
\name{oa_generate_records}
\alias{oa_generate_records}
\title{A generator for making request to OpenAlex API
TODO check group_by}
\usage{
oa_generate_records(
  query_url,
  per_page = 200,
  mailto = oa_email(),
  api_key = oa_apikey(),
  verbose = FALSE
)
}
\arguments{
\item{query_url}{Character string.
A search query formulated using the OpenAlex API language and
can be generated with \code{oa_query}.}

\item{per_page}{Numeric. Number of items to download per page.
The per-page argument can assume any number between 1 and 200.
Defaults to 200.}

\item{mailto}{Character string.
Gives OpenAlex an email to enter the polite pool.}

\item{api_key}{Character string.
Your OpenAlex Premium API key, if available.}

\item{verbose}{Logical.
If TRUE, print information about the querying process. Defaults to TRUE.}
}
\description{
A generator for making request to OpenAlex API
TODO check group_by
}
\examples{
\dontrun{
query_url <- "https://api.openalex.org/works?filter=cites\%3AW2755950973"
oar <- oa_generate_records(
  query_url, per_page = 10,
  verbose = TRUE
)
p1 <- oar() # item 1
p2 <- oar() # item 2
p3 <- oar() # item 3
head(p1)
head(p3)
}

}
