% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_ngram.R
\name{oa_ngrams}
\alias{oa_ngrams}
\title{Get N-grams of works}
\usage{
oa_ngrams(works_identifier, mailto = oa_email(), verbose = FALSE)
}
\arguments{
\item{works_identifier}{Character. OpenAlex ID(s) of "works" entities as item identifier(s).
These IDs normally start with "W". See more at <https://docs.openalex.org/about-the-data/work#id>.}

\item{mailto}{Character string.
Gives OpenAlex an email to enter the polite pool.}

\item{verbose}{Logical.
If TRUE, print information about the querying process. Defaults to TRUE.}
}
\value{
A dataframe of paper metadatada and a list-column of ngrams.
}
\description{
Some work entities in OpenAlex include N-grams (word sequences and their frequencies) of their full text.
The N-grams are obtained from Internet Archive, which uses the spaCy parser to index scholarly works.
See <https://docs.openalex.org/api/get-n-grams> for coverage and more technical details.
}
\examples{
\dontrun{

ngrams_data <- oa_ngrams(c("W1963991285", "W1964141474"))

library(dplyr)
first_paper_ngrams <- ngrams_data$ngrams[[1]]
top_10_ngrams <- first_paper_ngrams \%>\%
  slice_max(ngram_count, n = 10, with_ties = FALSE)

}
}
