% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaApiRequest.R
\name{oa_random}
\alias{oa_random}
\title{oa_fetch but for a random query}
\usage{
oa_random(
  entity = oa_entities(),
  output = c("tibble", "dataframe", "list"),
  endpoint = "https://api.openalex.org/"
)
}
\arguments{
\item{entity}{Character. It indicates the scholarly entity of the search.
The argument can be one of c("works", "authors", "venues", "institutions", "concepts").
If not provided, `entity` is guessed from `identifier`.}

\item{output}{Character. Type of output, either a list or a tibble/data.frame.}

\item{endpoint}{is character. It indicates the url of the OpenAlex Endpoint API server. The default value is endpoint = "https://api.openalex.org/".}
}
\value{
A data.frame or a list. One row or one element.
Result of the random query.
}
\description{
oa_fetch but for a random query
}
\examples{
oa_random()
}
