% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaInstitutions2df.R
\name{oaInstitutions2df}
\alias{oaInstitutions2df}
\title{Convert OpenAlex collection of institutions' records from list format to data frame}
\usage{
oaInstitutions2df(data, verbose = TRUE)
}
\arguments{
\item{data}{is a list. data is the output of the function \code{oaApiRequest}.}

\item{verbose}{is a logical. If TRUE, information about the querying process will be plotted on screen. Default is \code{verbose=TRUE}.}
}
\value{
a data.frame.

For more extensive information about OpenAlex API, please visit: \href{https://docs.openalex.org/api}{https://docs.openalex.org/api}
}
\description{
It converts bibliographic collection of institutions' records gathered from OpenAlex database \href{https://openalex.org/}{https://openalex.org/} into data frame.
The function converts a list of institutions' records obtained using \code{oaApiRequest} into a data frame/tibble.
}
\examples{
\dontrun{

# Query to search information about all Italian educational institutions


query_inst <- oaQueryBuild(
  entity = "institutions",
  filter = "country_code:it,type:education"
)

res <- oaApiRequest(
  query_url = query_inst,
  count_only = FALSE,
  verbose = FALSE
)

df <- oa2df(res, entity = "institutions")

df
}

# @export
}
