% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex_api.R
\name{openalex_api}
\alias{openalex_api}
\title{openalex_api}
\usage{
openalex_api(path)
}
\arguments{
\item{path}{the url path to use in the GET request to the openalex API}
}
\value{
list of parsed JSON returned from the API
}
\description{
Baseline function for calling the API and parsing results
}
\examples{
openalex_api(path = "https://api.openalex.org/works/W2741809807")
}
