% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa2df.R
\name{process_affil}
\alias{process_affil}
\title{Process affiliations}
\usage{
process_affil(l_institution)
}
\arguments{
\item{l_institution}{List. Nested elements include
id, display_name, ror, country_code, type, lineage}
}
\value{
Dataframe of with the following columns:
id, display_name, ror, country_code, type, lineage
}
\description{
Process affiliations
}
