% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa2df.R
\name{process_paper_authors}
\alias{process_paper_authors}
\title{Process paper authorships}
\usage{
process_paper_authors(authorships)
}
\arguments{
\item{authorships}{List. Authorships element of paper.}
}
\value{
List. A list of one dataframe with the processed authors:
id, display_name, orcid, author_position, is_corresponding, affiliations, affiliation_raw
}
\description{
Process paper authorships
}
\keyword{internal}
