% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_topics}
\alias{process_topics}
\title{Process topics}
\usage{
process_topics(entity, extra)
}
\arguments{
\item{entity}{List. One single work or author to process.}

\item{extra}{Character. Either "score" (work) or "count" (author).}
}
\value{
List. A list of one tibble with the processed topics.
}
\description{
Process topics
}
\keyword{internal}
