% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_snowball.R
\name{to_disk}
\alias{to_disk}
\title{Flatten snowball result}
\usage{
to_disk(snowball)
}
\arguments{
\item{snowball}{List result from `oa_snowball`.}
}
\value{
Tibble/data.frame of works with additional `cited_by` column.
}
\description{
|  id|title |...|cited_by_count| referenced_works   |cited_by |...|
| 100|foo   |...|             1| 98, 99             |101      |...|
| 200|bar   |...|             2| 198, 199           |201, 202 |...|
| 300|wug   |...|             2| 296, 297, 298, 299 |301, 302 |...|
}
\examples{
\dontrun{
flat_snow <- to_disk(oa_snowball(
  identifier = "W1516819724",
  verbose = TRUE
))

flat_snow[, c("id", "cited_by")]
}
}
