% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode.R
\name{countrycode}
\alias{countrycode}
\title{Index of Countries and their alpha-2 and aplha-3 codes.}
\format{
A data frame with 249 rows and 3 variables:
 \describe{
    \item{Country}{country names}
    \item{Alpha2}{countries' alpha-2 codes}
    \item{Alpha3}{countries' alpha-3 codes}
    }
}
\description{
Data frame contains the list of countries and their alpha-2 and aplha-3 codes.\cr
}
