% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundational_functions.R
\name{get_number_of_pages}
\alias{get_number_of_pages}
\title{Get Number of Pages}
\usage{
get_number_of_pages(author_papers)
}
\arguments{
\item{author_papers}{}
}
\value{

}
\description{
Get Number of Pages
}
