% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa2df.R
\name{oa2df}
\alias{oa2df}
\title{Convert OpenAlex collection from list to data frame}
\usage{
oa2df(data, entity, count_only = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{is a list. data is the output of the function \code{oa_request}.}

\item{entity}{is a character. It indicates the scholarly entity of the search. The argument can be equal to
entity = c("works", "authors", "venues", "institutions", "concepts"). The default value is entity = works".}

\item{count_only}{Logical. If TRUE, the function returns only the number of item matching the query. Default is \code{count_only=FALSE}.}

\item{verbose}{is a logical. If TRUE, information about the querying process will be plotted on screen. Default is \code{verbose=TRUE}.}
}
\value{
a data.frame.

For more extensive information about OpenAlex API, please visit: \href{https://docs.openalex.org/api}{https://docs.openalex.org/api}
}
\description{
It converts bibliographic collections gathered from OpenAlex database \href{https://openalex.org/}{https://openalex.org/} into data frame.
The function converts a collection of records about works, authors, institutions, venues or concepts obtained using \code{oa_request} into a data frame/tibble.
}
\examples{
\dontrun{

# Query to search all works citing the article:
#  Aria, M., & Cuccurullo, C. (2017). bibliometrix:
#   An R-tool for comprehensive science mapping analysis.
#   Journal of informetrics, 11(4), 959-975.

#  published in 2021.
#  The paper is associated to the OpenAlex id W2755950973.

#  Results have to be sorted by relevance score in a descending order.

query <- oa_query(
  entity = "works",
  cites = "W2755950973",
  from_publication_date = "2021-01-01",
  to_publication_date = "2021-12-31",
  endpoint = "https://api.openalex.org/"
)

res <- oa_request(
  query_url = query,
  count_only = FALSE,
  verbose = FALSE
)

df <- oa2df(res, entity = "works")

df
}

}
