% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa2df.R
\name{oa2df}
\alias{oa2df}
\title{Convert OpenAlex collection from list to data frame}
\usage{
oa2df(data, entity = "works")
}
\arguments{
\item{data}{is a list. data is the output of the function \code{oaApiRequest}.}
}
\value{
a data.frame.

For more extensive information about OpenAlex API, please visit: \href{https://docs.openalex.org/api}{https://docs.openalex.org/api}
}
\description{
It converts bibliographic collection gathered from OpenAlex database \href{https://openalex.org/}{https://openalex.org/} into data frame.
The function converts a collection of records about works, authors, institutions, venues or concepts obtained using \code{oaApiRequest} into a data frame/tibble.
}
