% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coro.R
\name{oa_generate}
\alias{oa_generate}
\title{A generator for making request to OpenAlex API
Returns one record at a time.}
\usage{
oa_generate(
  query_url,
  mailto = oa_email(),
  api_key = oa_apikey(),
  verbose = FALSE
)
}
\arguments{
\item{query_url}{Character string.
A search query formulated using the OpenAlex API language and
can be generated with \code{oa_query}.}

\item{mailto}{Character string.
Gives OpenAlex an email to enter the polite pool.}

\item{api_key}{Character string.
Your OpenAlex Premium API key, if available.}

\item{verbose}{Logical.
If TRUE, print information about the querying process. Defaults to TRUE.}
}
\value{
Generator function.
}
\description{
TODO check group_by
}
\examples{
if (require("coro")) {
  query_url <- "https://api.openalex.org/works?filter=cites\%3AW2755950973"
  oar <- oa_generate(query_url, verbose = TRUE)
  print(oar)
  p1 <- oar() # record 1
  p2 <- oar() # record 2
  p3 <- oar() # record 3
  head(p1)
  head(p3)
}

}
