% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coro.R
\name{oa_generate}
\alias{oa_generate}
\title{A generator for making request to OpenAlex API
Returns one record at a time.}
\usage{
oa_generate(...)
}
\arguments{
\item{...}{arguments passed to the generator including
`query_url`, `mailto`, `api_key`, and `verbose`.
See `oa_request` for details on these arguments.}
}
\value{
Generator function.
}
\description{
TODO check group_by
}
\examples{
if (require("coro")) {
  query_url <- "https://api.openalex.org/works?filter=cites\%3AW2755950973"
  oar <- oa_generate(query_url, verbose = TRUE)
  p1 <- oar() # record 1
  p2 <- oar() # record 2
  p3 <- oar() # record 3
  head(p1)
  head(p3)
}

}
