% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_fetch.R
\name{oa_random}
\alias{oa_random}
\title{oa_fetch but for a random query}
\usage{
oa_random(
  entity = oa_entities(),
  output = c("tibble", "dataframe", "list"),
  endpoint = "https://api.openalex.org"
)
}
\arguments{
\item{entity}{Character. Scholarly entity of the search.
The argument can be one of
c("works", "authors", "institutions", "keywords", "funders", "sources", "publishers", "topics").
If not provided, `entity` is guessed from `identifier`.}

\item{output}{Character.
Type of output, one of `"tibble"`, `"dataframe"`, `"list"`, or `"raw"`.
\describe{
 \item{tibble}{a tibble tidy data}
 \item{dataframe}{a base data.frame tidy data}
 \item{list}{a list of parsed JSON contents}
 \item{raw}{a list of raw JSON strings (length depends on query)}
}}

\item{endpoint}{Character. URL of the OpenAlex Endpoint API server.
Defaults to endpoint = "https://api.openalex.org".}
}
\value{
A data.frame or a list. One row or one element.
Result of the random query.
If you would like to select more than one random entity, say, 10,
use `options = list(sample = 10)` argument in `oa_fetch`.
}
\description{
oa_fetch but for a random query
}
\examples{
oa_random()
}
