% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_snowball.R
\name{oa_snowball}
\alias{oa_snowball}
\title{A function to perform a snowball search
and convert the result to a tibble/data frame.}
\usage{
oa_snowball(
  identifier = NULL,
  ...,
  id_type = c("short", "original"),
  options = list(),
  mailto = oa_email(),
  endpoint = "https://api.openalex.org",
  verbose = FALSE,
  citing_filter = list(),
  cited_by_filter = list()
)
}
\arguments{
\item{identifier}{Character vector of openalex_id identifiers.}

\item{...}{Additional arguments to pass to `oa_fetch` when querying the
input works, such as `doi`.}

\item{id_type}{Type of OpenAlex IDs to return. Defaults to "short",
which remove the prefix https://openalex.org/ in the works' IDs,
for example, W2755950973.
If "original", the OpenAlex IDs are kept as are,
for example, https://openalex.org/W2755950973}

\item{options}{List. Additional parameters to add in the query. For example:
- `select` Character vector. Top-level fields to show in output.
Defaults to NULL, which returns all fields.
https://docs.openalex.org/how-to-use-the-api/get-single-entities/select-fields
- `sort` Character. Attribute to sort by.
For example: "display_name" for venues or "cited_by_count:desc" for works.
See more at <https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/sort-entity-lists>.
- `sample` Integer. Number of (random) records to return.
Should be no larger than 10,000.
Defaults to NULL, which returns all records satisfying the query.
Read more at <https://docs.openalex.org/how-to-use-the-api/get-lists-of-entities/sample-entity-lists>.
- `seed` Integer.
A seed value in order to retrieve the same set of random records in
the same order when used multiple times with `sample`.
IMPORTANT NOTE: Depending on your query, random results with a seed value may change over time due to new records coming into OpenAlex.
This argument is likely only useful when queries happen close together (within a day).}

\item{mailto}{Character string.
Gives OpenAlex an email to enter the polite pool.}

\item{endpoint}{Character. URL of the OpenAlex Endpoint API server.
Defaults to endpoint = "https://api.openalex.org".}

\item{verbose}{Logical. If TRUE, print information on querying process.
Default to \code{verbose = FALSE}.
To display the full query URL, set the environment variable openalexR.debug to TRUE.}

\item{citing_filter}{filters used in the search of works citing the input works.}

\item{cited_by_filter}{filters used in the search of works cited by the input works.}
}
\value{
A list containing 2 elements:
- nodes: dataframe with publication records.
The last column `oa_input` indicates whether the work was
one of the input `identifier`(s).
- edges: publication link dataframe of 2 columns `from, to`
such that a row `A, B` means A -> B means A cites B.
In bibliometrics, the "citation action" comes from A to B.
}
\description{
A function to perform a snowball search
and convert the result to a tibble/data frame.
}
\examples{
\dontrun{

snowball_docs <- oa_snowball(
  identifier = c("W2741809807", "W2755950973"),
  citing_filter = list(from_publication_date = "2022-01-01"),
  cited_by_filter = list(),
  verbose = TRUE
)
}
}
