% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaApiRequest.R
\name{oaApiRequest}
\alias{oaApiRequest}
\title{Get bibliographic records from OpenAlex databases
It gets bibliographic records from OpenAlex database \href{https://openalex.org/}{https://openalex.org/}.
The function \code{oaApiRequest} queries OpenAlex database using a query formulated through the function \code{oaQueryBuild}.}
\usage{
oaApiRequest(
  query_url,
  format = "data.frame",
  overview = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{query_url}{is a character. It contains a search query formulated using the OpenAlex API language. A query can be automatically generated using the function \code{oaQueryBuild}.}

\item{format}{is a character. It indicates the data format of the returned object. The argument can be \code{format=c("data.frame","list")}. Default is \code{format="data.frame"}.}

\item{verbose}{is a logical. If TRUE, information about the querying process will be plotted on screen. Default is \code{verbose=FALSE}.}
}
\value{
a data.frame or a list.

For more extensive information about OpenAlex API, please visit: \href{https://docs.openalex.org/api}{https://docs.openalex.org/api}
}
\description{
Get bibliographic records from OpenAlex databases
It gets bibliographic records from OpenAlex database \href{https://openalex.org/}{https://openalex.org/}.
The function \code{oaApiRequest} queries OpenAlex database using a query formulated through the function \code{oaQueryBuild}.
}
\examples{

\dontrun{
# Query to search all works citing the article:
  Aria, M., & Cuccurullo, C. (2017). bibliometrix:
  An R-tool for comprehensive science mapping analysis.
  Journal of informetrics, 11(4), 959-975.

  published in 2021.
  The paper is associated to the OpenAlex id W2755950973.

  Results have to be sorted by relevance score in a descending order.

query <- oaQueryBuild(
identifier=NULL,
entity = "works",
filter="cites:W2755950973,publication_year:2021",
search=NULL,
sort="relevance_score:desc",
endpoint = "https://api.openalex.org/")

res <- oaApiRequest(query,
               format = "data.frame",
               verbose=T)
}


}
