% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalexR-package.R
\docType{package}
\name{openalexR-package}
\alias{openalexR}
\alias{openalexR-package}
\title{openalexR: Getting Bibliographic Records from 'OpenAlex' Database Using 'DSL' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools to extract bibliographic content from 'OpenAlex' database using API \url{https://docs.openalex.org/api/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/massimoaria/openalexR}
  \item \url{https://massimoaria.github.io/openalexR/}
  \item Report bugs at \url{https://github.com/massimoaria/openalexR/issues}
}

}
\author{
\strong{Maintainer}: Massimo Aria \email{massimo.aria@gmail.com} (\href{https://orcid.org/0000-0002-8517-9411}{ORCID}) [copyright holder]

}
\keyword{internal}
